Adds minimum version requirement for kubectl (#8415)
* saving untested changes

* Fix for issue #8341

* Fix tab navigation within modal dialog (#8326)

* Fix tab navigation within modal dialog

* Add import

* Fix spelling

* Change to just add/remove items from DOM as necessary

* Fix a couple a11y issues with manage access (#8386)

* Fix a couple a11y issues with manage access

* Fix strict null check

* Fix another strict null check

* Update js file with monaco -> ads workbench (#8398)

* Set encoding to true (Bug fix for flavor) (#8395)

For offline scripts it shows the flavor as “Choose SQL Language”, this is because in flavorStatus.ts line 150 when we compare (uri === currentUri), the value of current Uri is "file:///d%3A/GitHub/PGExtension/TestDatabase/1ae730a9.sql" whereas the value for uri is "file:///d:/GitHub/PGExtension/TestDatabase/1ae730a9.sql" which ends up returning label ‘Choose SQL Language’.

In queryInput.ts we set public get uri(): string { return this.getResource().toString(true); }
This is the variable that we use in doChangeLanguageFlavor. And as we compare this with getEditorUri() later in flavorStatus.ts, it does not match. So enabling encoding here as well to get the encoded string in both the cases.

* bump handlebars in extensions (#8411)

* remove active css class on mouse leave (#8410)

* Bump SqlToolsService (#8404)

* Remove open notebook entry points (#8393)

* Add aria role and selected properties (#8405)

* Add aria role and selected properties

* Add img role fix

* Add title to text

* Notebooks cleanup: Remove old out of proc markdown option (#8394)

* Remove old out of proc markdown option

* Revert change to stats.js

* remove debug console.logs

* PR feedback