Update extension to .NET 6.
This change updates the extension reference to .NET 6, instead of .NET Core. This should partially address Charles' concern about havin extensions with short lifespan. We agreed that we will release it scoped to .NET 6 now and when .NET 8 (next LTS release) comes out, we can revisit this and see if we want to consolidate them, based on the amount of breaking changes introduced in version 8.

Ialso updated flags to `hidden`, as suggested by Karl and added link to the changelog.