Add .NET Core 3.1 Runtime extension to the marketplace. (#17094)
* Add .NET Core 3.1 Runtime extension to the marketplace.

This is the supporting extension, on which all other extensions that we build will depend. The extension by itself does not add any functionality to ADS and merely allows other extensions to use .NET Core runtime without requiring user to install it manually.

* Add source code URL.

Given that we do not publish source code for this extension just yet, the best we can do is to point to the Microsoft organization at github. We will update this link once we go through the OSS process and publish the source code for this extension.

* Update extension to .NET 6.

This change updates the extension reference to .NET 6, instead of .NET Core. This should partially address Charles' concern about havin extensions with short lifespan. We agreed that we will release it scoped to .NET 6 now and when .NET 8 (next LTS release) comes out, we can revisit this and see if we want to consolidate them, based on the amount of breaking changes introduced in version 8.

Ialso updated flags to `hidden`, as suggested by Karl and added link to the changelog.

Co-authored-by: Charles Gagnon <chgagnon@microsoft.com>