Set role to button for the wizard step links.
Our extension is relying on the wizard dialog. During accessibility testing it was discovered that wizard step buttons are being reported as links by the screen reader (NVDA, JAWS). Claimed expected behavior by the tester is that they should be announced as buttons.

This change addresses the problem by adding `role` attribute to the said elements with the value of a `button`.