Respect ARIA label specified int he tree component options.
During accessiblity testing it was discovered that tree view in our wizard reads "Tree Node treeview" instead of the proper label that is specified. It turned out to be the problem with the tree component, where `ariaLabel` was hardcoded to "Tree Node", instead of the one provided inthe options.

This change addresses the problem by passing throught `ariaLabel` from the options object to the underlying tree control. I also removed the default `Tree Node` hardocded label, as it didn't make much sense. This does mean that all tree-views that do not explicilty specify their aria-label will now get an empty label. I think this is better than having unrelated, unlocalized `Tree Node`.

I'm also worried about changes to the `ariaLabel` property after the component was initialized. I updated the code to propagate the value to the underlying treeview in the `setProperties` override of the tree component and hope that it will take care of it.