Preserve Azure account for any auth type
When a service has been selected through the Azure
browser, we want to preserve the Azure account
information even if a different authentication has
been selected. This allows doing ARM operations
using the signed in Azure account for any resources
including those with a different login types such
as clusters.